+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 # 1.1.0 [2019-05-10]
+- src/static folder for static resources.
+- remake the `useSocket` hook, it is now a single global socket.
 
 # 1.0.0 [2019-05-10]
 - Initial Version